Fix opcode reading

+++ src/interpretor.rs
@@ -1,14 +1,14 @@
 use crate::display::Display;
 
 pub struct Interpretor {
-    program: Vec<u8>,
+    program: Vec<u16>,
     screen: Display,
     registers: [u8; 16],
     l: u16
 }
 
 impl Interpretor {
-	pub fn new(program: Vec<u8>) -> Interpretor {
+	pub fn new(program: Vec<u16>) -> Interpretor {
 		Interpretor {
 			program: program,
 			screen: Display::new(64, 32),
@@ -17,7 +17,7 @@ impl Interpretor {
 		}
 	}
 
-	fn decode(&mut self, op: u8) {
+	fn decode(&mut self, op: u16) {
 		match op {
 			0x0000 => {}, // Ignored
 			0x00E0 => { self.screen.clear() },
